refactor(auth): 인증 관련 유틸 함수 명칭 수정
- token.ts -> auth.ts

Related to #29
