+++ src/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './date';
 export * from './color';
-export * from './token';
+export * from './auth';
